Changes to prepare the GUI_alignment branch for merging into main branch. This includes rectangular region selection and how parameters are reset when selecting the reset button.
