+++ GUI/align.py
@@ -10,7 +10,7 @@ Tim Cote, ANL, Fall 2019.
 
 import os
 from os import path
-from util import join, flatten_order_list
+from util import join, flatten_order_list, check_setup, read_fls, pull_image_files
 from typing import Any, Dict, List, Optional, Tuple, Union
 
 # ============================================================= #
@@ -63,222 +63,6 @@ def format_macro(all_macros: List[str]) -> str:
     joined_macro = join(all_macros, '\n')
     full_macro = fijify_macro(joined_macro)
     return full_macro
-
-
-# ============================================================= #
-#                    Manipulating FLS Files                     #
-# ============================================================= #
-def pull_image_files(fls_file: str,
-                     check_align: bool = False) -> List[List[str]]:
-    """Use .fls file to return ordered images for alignment.
-
-    Initially it will read in .fls data, pull the number of files
-    from the first line, and then locate the in-focus image. Then
-    it separates the overfocus and underfocus images.
-
-    If the check alignment is set, the returned images are the infocus image,
-    and the nearest under-focused/over-focused on either side of the
-    infocus image. Otherwise all image files are returned. The files are
-    ordered from
-
-
-                    [
-                    [smallest underfocus,  ... , largest underfocus]
-                    [infocus]
-                    [smallest overfocus, ... , largest overfocus]
-                    ]
-
-    Args:
-        fls_file: The filename for the fls file.
-        check_align: Option for full alignment or parameter test.
-
-    Returns:
-        filenames: 2D list of under/in/over-focus images.
-    """
-
-    # Read data.
-    with open(fls_file) as fls_text:
-        fls_text = fls_text.read()
-
-    # Find focus line demarcations in .fls file
-    fls_lines = fls_text.splitlines()
-    num_files = int(fls_lines[0])
-    under_split = 1
-    focus_split = num_files // 2 + under_split
-    over_split = num_files // 2 + focus_split
-
-    # Grab infocus image and defocus images.
-    # If checking parameters, only grab one
-    # file on either side of the infocus image.
-    focus_file = [fls_lines[focus_split]]
-    if check_align:
-        under_files = [fls_lines[focus_split - 1]]
-        over_files = [fls_lines[focus_split + 1]]
-    else:
-        under_files = fls_lines[under_split:focus_split]
-        over_files = fls_lines[focus_split + 1: over_split + 1]
-
-    # Reverse underfocus files due to how ImageJ opens images
-    filenames = [under_files[::-1], focus_file, over_files]
-    return filenames
-
-
-def grab_fls_data(fls1: str, fls2: str, tfs_value: str,
-                  fls_value: str, check_sift: bool) -> Tuple[List[str], List[str]]:
-    """Grab image data from .fls file.
-
-    Given the FLS files for the flip/unflip/single images,
-    return the image filenames depending on the fls_value and
-    through-focal-series (tfs) value.
-
-    Examples:
-        - 1 FLS, Unflip/FLip -> files1 : populated, files2 : populated
-        - 1 FLS, Single      -> files1 : populated, files2 : empty
-        - 2 FLS, Unflip/FLip -> files1 : populated, files2 : populated
-
-    Args:
-        fls1: Filename for 1st FLS file.
-        fls2: Filename for 2nd FLS file.
-        tfs_value: Value for type of through-focal series,
-            Single or Unflip/Flip.
-        fls_value: String of number of FLS files used.
-        check_sift: Option to check sift alignment.
-
-    Returns:
-        Tuple of image filenames for scenarios above
-            - files1: List of image filenames
-                according to examples above.
-            - files2: List of image filenames.
-                according to examples above.
-    """
-
-    # Read image data from .fls files and store in flip/unflip lists
-    if fls_value == 'One':
-        files1 = pull_image_files(fls1, check_sift)
-        if tfs_value == 'Unflip/Flip':
-            files2 = pull_image_files(fls1, check_sift)
-        else:
-            files2 = []
-    elif fls_value == 'Two':
-        files1 = pull_image_files(fls1, check_sift)
-        files2 = pull_image_files(fls2, check_sift)
-    return files1, files2
-
-
-def read_fls(path1: Optional[str], path2: Optional[str], fls_files: List[str],
-             tfs_value: str, fls_value: str,
-             check_sift: bool = False) -> Tuple[List[str], List[str]]:
-    """Read image files from .fls files and returns their paths.
-
-    The images are read from the FLS files and the files are returned
-    depending on the through-focal series value and the fls value. Once
-    image filenames are pulled from the FLS file, they are joined to
-    the paths (directories) the images are stored in. Those resulting
-    full path names are returned in files1 and files2.
-
-    Args:
-        path1: The first unflip/flip/single path/directory. Optional
-        path2: The first unflip/flip/single path/directory. Optional
-        fls_files: A list of the FLS filenames.
-        tfs_value: The through-focal series option.
-                Options: Unflip/FLip, Single
-        fls_value: The FLS option.
-                Options: One, Two
-        check_sift: Option for checking SIFT alignment.
-
-    Returns:
-        (files1, files2): A tuple of the lists of image paths corresponding
-            to path1 and path2.
-            - files1: List of full image paths.
-            - files2: List of full image paths or empty list.
-    """
-
-    # Find paths and .fls files
-    fls1, fls2 =  fls_files[0], fls_files[1]
-
-    # Read image data from .fls files and store in flip/unflip lists
-    files1, files2 = grab_fls_data(fls1, fls2, tfs_value, fls_value, check_sift)
-
-    print('image names:', files1, files2)
-
-    # Check same number of files between fls
-    if tfs_value != 'Single':
-        if len(flatten_order_list(files1)) != len(flatten_order_list(files2)):
-            return
-    # Check if image path exists and break if any path is nonexistent
-    if path1 is None and path2 is None:
-        return
-    for file in flatten_order_list(files1):
-        full_path = join([path1, file], '/')
-        if not path.exists(full_path):
-            print(full_path, " doesn't exist!")
-            return
-    if files2:
-        for file in flatten_order_list(files2):
-            full_path = join([path2, file], '/')
-            if not path.exists(full_path):
-                print(full_path, " doesn't exist!")
-                return
-    return files1, files2
-
-
-def check_setup(datafolder: str, tfs_value: str,
-                fls_value: str, fls_files: List[str],
-                prefix: str = '') -> Check_Setup:
-    """Check to see all images filenames in .fls exist in datafolder.
-
-    Args:
-        datafolder: Datafolder path.
-        tfs_value: The through-focal series option.
-                Options: Unflip/FLip, Single
-        fls_value: The FLS option.
-                Options: One, Two
-        fls_files: A list of the FLS filenames.
-        prefix: The prefix to prepend for print statements in GUI.
-
-    Returns:
-        vals: Will return images filenames and paths to those files
-            and their parent directories if all images pulled from FLS exist.
-            - vals[0]: Corresponds to process success.
-            - vals[1]: 1st parent directory path or None.
-            - vals[2]: 2nd parent directory path or None.
-            - vals[3]: 1st path list of ordered image filenames.
-            - vals[4]: 2nd path list of ordered image filenames.
-    """
-
-    # Find paths and .fls files
-    if tfs_value == 'Unflip/Flip':
-        path1 = join([datafolder, 'unflip'], '/')
-        path2 = join([datafolder, 'flip'], '/')
-    elif tfs_value == 'Single':
-        path1 = join([datafolder, 'tfs'], '/')
-        if not os.path.exists(path1):
-            path1 = join([datafolder, 'unflip'], '/')
-            if not os.path.exists(path1):
-                path1 = None
-        path2 = None
-
-    # Grab the files that exist in the flip and unflip dirs.
-    file_result = read_fls(path1, path2, fls_files,
-                           tfs_value, fls_value, check_sift=False)
-
-    print('file_result', file_result)
-
-    vals = (False, None, None, None, None)
-    if isinstance(file_result, tuple):
-        files1, files2 = file_result
-        flattened_files1 = flatten_order_list(files1)
-        flattened_files2 = None
-        if files2:
-            flattened_files2 = flatten_order_list(files2)
-        vals = (True, path1, path2, flattened_files1, flattened_files2)
-    # Prints if task failed
-    else:
-        print(f'{prefix}Task failed because the number of files extracted from the directory', end=' ')
-        print(f'does not match the number of files expected from the .fls file.')
-        print(f'{prefix}Check that filenames in the flip, unflip, or tfs', end=' ')
-        print(f'path match and all files exist in the right directories.')
-    return vals
 
 
 # ============================================================= #
