+++ GUI/gui_styling.py
@@ -10,6 +10,7 @@ Timothy Cote, ANL, Fall 2019.
 import os
 from typing import Any, Dict, Tuple, Optional
 
+
 # ============================================================= #
 #                          Window Scaling                       #
 # ============================================================= #
@@ -89,8 +90,6 @@ class WindowStyle(object):
         # Input size
         inp_pad = ((0, 0), (0, 10))
         inp_wid = (5, 1)
-        inp_d = {'pad': inp_pad,
-                 'size': inp_wid}
 
         # Fonts
         T_f = self.fonts["title"]
@@ -116,25 +115,14 @@ class WindowStyle(object):
 
             # __________________ Home  ______________ #
             # Browse Button elements
-            if key == '__Fiji_Browse__':
-                val.update(pad=pad(164, 0, 10, 0), target='__Fiji_Path__')
-            elif key == '__Browser_Browse__':
+            if key == '__Browser_Browse__':
                 val.update(pad=pad(164, 0, 10, 0), target='__Browser_Path__')
             # Button elements
             elif key == '__Browser_Set__':
                 val.update(pad=pad(10, 0, 10, 0))
             elif key == '__Browser_Reset__':
                 val.update(pad=pad(10, 0, 10, 0))
-            elif key == '__Fiji_Set__':
-                val.update(pad=pad(53, 0, 10, 0))
-            elif key == '__Fiji_Reset__':
-                val.update(pad=pad(10, 0, 10, 0))
             # Input elements
-            elif key == '__Fiji_Path__':
-                val.update(pad=pad(16, 0, 10, 0), size=(30, 1), enable_events=True,
-                           metadata={'Set': '', 'Def': '', 'State': 'Def'},
-                           use_readonly_for_disable=True,
-                           disabled_readonly_background_color='#A7A7A7')
             elif key == '__Browser_Path__':
                 val.update(pad=pad(16, 0, 10, 0), size=(30, 1))
             # Text elements
@@ -149,442 +137,6 @@ class WindowStyle(object):
                            font=B_f, background_color='#e4eaf5')
             elif key == 'home_contact':
                 val.update(pad=pad(235, 0, 20, 40))
-            return val
-
-        def LS_style(key, val) -> Dict:
-            """The styles within the Linear Sift tab of the GUI.
-
-            Args:
-                key: The key for a PySimpleGUI element.
-                val: The dictionary of the style with which to update key.
-
-            Returns:
-                key_style: A dictionary of the key-value pairs
-                    for a specific type of element given
-                    by 'key'."""
-
-            # _______________ Linear Sift Tab _______________ #
-            # Invisible
-            if key == '__LS_FLS1_Staging__':
-                val.update(visible=False, enable_events=True)
-            elif key == '__LS_FLS2_Staging__':
-                val.update(visible=False, enable_events=True)
-            # Browse buttons element
-            elif key == '__LS_Image_Dir_Browse__':
-                val.update(pad=pad(8, 0, 0, 0), initial_folder=default_folder,
-                           target='__LS_Image_Dir_Path__')
-            elif key == '__LS_Load_FLS1__':
-                val.update(target='__LS_FLS1_Staging__', pad=((10, 0), (5, 0)),
-                           disabled=True)
-            elif key == '__LS_Load_FLS2__':
-                val.update(target='__LS_FLS2_Staging__', pad=((10, 0), (5, 0)),
-                           disabled=True)
-            # Button element
-            elif key == '__LS_Set_Img_Dir__':
-                val.update(pad=pad(5, 0, 0, 0),
-                           metadata={'State': 'Def'})
-            elif key == '__LS_Reset_Img_Dir__':
-                val.update(pad=pad(5, 0, 0, 0))
-            elif key == '__LS_Run_Align__':
-                val.update(pad=pad(50, 0, 5, 0), disabled=True)
-            elif key == '__LS_Adjust__':
-                val.update(pad=pad(20, 0, 0, 0), disabled=True,
-                           metadata={'Set': 'Quit Adjustment', 'Def': 'Adjust Image',
-                                     'State': 'Def'})
-            elif key == '__LS_View_Stack__':
-                val.update(pad=pad(185, 0, 0, 0), disabled=True,
-                           metadata={'Set': 'Hide Stack', 'Def': 'View Stack',
-                                     'State': 'Def'})
-            elif key == '__LS_Set_FLS__':
-                val.update(pad=((116, 0), (4, 5)), disabled=True,
-                           metadata={'State': 'Def'})
-            elif key == '__LS_Reset_FLS__':
-                val.update(pad=((10, 116), (4, 5)), disabled=True)
-            # Checkbox element
-            elif key == '__LS_horizontal_flip__':
-                val.update(default=True, pad=pad(35, 0, 0, 0), disabled=True)
-            elif key == '__LS_interp__':
-                val.update(pad=pad(200, 0, 4, 20), default=True)
-            # Combo element
-            elif key == '__LS_exp_transf__':
-                val.update(pad=pad(0, 10, 4, 10), size=(9, 1),
-                           default_value='Affine', readonly=True)
-            elif key == '__LS_FLS_Combo__':
-                val.update(enable_events=True, default_value='Two', size=(5, 1),
-                           readonly=True, disabled=True, pad=((3, 0), (10, 0)),
-                           metadata={'Set': 'Two', 'Def': 'Two', 'State': 'Def'})
-            elif key == '__LS_TFS_Combo__':
-                val.update(enable_events=True, default_value='Unflip/Flip', size=(11, 1),
-                           readonly=True, disabled=True, pad=((0, 0), (10, 0)),
-                           metadata={'Set': 'Unflip/Flip', 'Def': 'Unflip/Flip',
-                                     'State': 'Def'})
-            # Graph element
-            elif key == '__LS_Graph__':
-                val.update(background_color='#808080', pad=pad(14, 14, 10, 0))
-
-            # Image elements
-            elif key == '__LS_Align_Spinner__':
-                val.update(filename=background_img, background_color=theme_background, pad=((5, 5), (5, 0)),
-                           metadata={'Set': spinner_gif, 'Def': background_img,
-                                     'State': 'Def'})
-            # Input field element
-            elif key == '__LS_Image_Dir_Path__':
-                val.update(size=(30, 1),
-                           use_readonly_for_disable=True,
-                           disabled_readonly_background_color='#A7A7A7')
-            elif key == '__LS_FLS1__':
-                val.update(enable_events=True, use_readonly_for_disable=True,
-                           disabled=True, size=(20, 1), pad=((10, 0), (5, 0)),
-                           disabled_readonly_background_color='#A7A7A7',
-                           metadata={'Set': 'None', 'Def': 'None', 'State': 'Def'})
-            elif key == '__LS_FLS2__':
-                val.update(enable_events=True, use_readonly_for_disable=True,
-                           disabled=True, size=(20, 1), pad=((10, 0), (5, 0)),
-                           disabled_readonly_background_color='#A7A7A7',
-                           metadata={'Set': 'None', 'Def': 'None', 'State': 'Def'})
-            elif key == '__LS_Image1__':
-                val.update(default_text='None', pad=pad(0, 0, 0, 0), enable_events=True,
-                           size=(30, 1), background_color=self.DEF_BACKGROUND,
-                           metadata={'Set': 'None', 'Def': 'None',
-                                     'State': 'Def'})
-            elif key == '__LS_Image2__':
-                val.update(default_text='None', pad=pad(0, 0, 0, 0),
-                           size=(30, 1), background_color=self.DEF_BACKGROUND, enable_events=True,
-                           metadata={'Set': 'None', 'Def': 'None',
-                                     'State': 'Def'})
-            elif key == '__LS_Stack__':
-                val.update(default_text='None', enable_events=True, pad=pad(19, 0, 0, 5),
-                           size=(30, 1), background_color=self.DEF_BACKGROUND,
-                           metadata={'Set': 'None', 'Def': 'None',
-                                     'State': 'Def'})
-            elif key == '__LS_igb__':
-                val.update(**inp_d)
-            elif key == '__LS_spso__':
-                val.update(**inp_d)
-            elif key == '__LS_min_im__':
-                val.update(**inp_d)
-            elif key == '__LS_max_im__':
-                val.update(**inp_d)
-            elif key == '__LS_fds__':
-                val.update(**inp_d)
-            elif key == '__LS_fdob__':
-                val.update(**inp_d)
-            elif key == '__LS_cncr__':
-                val.update(**inp_d)
-            elif key == '__LS_max_al_err__':
-                val.update(**inp_d)
-            elif key == '__LS_inlier_rat__':
-                val.update(**inp_d)
-            elif key == '__LS_transform_x__':
-                val.update(pad=(5, 5), size=(4, 1),
-                           use_readonly_for_disable=True, disabled=True,
-                           disabled_readonly_background_color='#A7A7A7')
-            elif key == '__LS_transform_y__':
-                val.update(pad=(5, 5), size=(4, 1),
-                           use_readonly_for_disable=True, disabled=True,
-                           disabled_readonly_background_color='#A7A7A7')
-            elif key == '__LS_transform_rot__':
-                val.update(pad=(5, 5), size=(4, 1), disabled=True,
-                           use_readonly_for_disable=True,
-                           disabled_readonly_background_color='#A7A7A7')
-            # Radio element
-            elif key == '__LS_full_align__':
-                val.update(default=True, size=(12, 2))
-            elif key == '__LS_param_test__':
-                val.update(size=(22, 2))
-            elif key == '__LS_unflip_reference__':
-                val.update(size=(12, 0), default=True, disabled=True, enable_events=True, pad=pad(76, 0, 0, 2))
-            elif key == '__LS_flip_reference__':
-                val.update(size=(12, 0), disabled=True, enable_events=True, pad=pad(10, 0, 0, 2))
-            # Slider element
-            elif key == '__LS_Stack_Slider__':
-                val.update(range=(0, 0), size=(30, 15), pad=pad(25, 0, 4, 4), default_value=0,
-                           enable_events=True, resolution=1, orientation='h', disable_number_display=True,
-                           metadata={'slider_range': (0, 0)})
-            # Text element
-            elif key == '__LS_FLS1_Text__':
-                val.update(pad=((5, 0), (5, 0)),
-                           metadata={'Both': 'Both FLS', 'Single': '1 FLS',
-                                     'Two': "Unflip FLS"})
-            elif key == '__LS_FLS2_Text__':
-                val.update(pad=((5, 0), (5, 0)),
-                           metadata={'Both': '', 'Single': '',
-                                     'Two': "Flip FLS"})
-            return val
-
-        def BUJ_style(key, val) -> Dict:
-            """The styles within the bUnwarpJ tab of the GUI.
-
-            Args:
-                key: The key for a PySimpleGUI element.
-                val: The dictionary of the style with which to update key.
-
-            Returns:
-                key_style: A dictionary of the key-value pairs
-                    for a specific type of element given
-                    by 'key'."""
-
-            # _______________ bUnwarpJ Tab _______________ #
-            # Invisible
-            if key == '__BUJ_Unflip_Stage_Load__':
-                val.update(visible=False, enable_events=True)
-            if key == '__BUJ_Flip_Stage_Load__':
-                val.update(visible=False, enable_events=True)
-            if key == '__BUJ_Mask_Stage_Load__':
-                val.update(visible=False, enable_events=True)
-            if key == '__BUJ_FLS1_Staging__':
-                val.update(visible=False, enable_events=True)
-            elif key == '__BUJ_FLS2_Staging__':
-                val.update(visible=False, enable_events=True)
-            # Browse buttons element
-            elif key == '__BUJ_Image_Dir_Browse__':
-                val.update(pad=pad(8, 0, 0, 0), target='__BUJ_Image_Dir_Path__', initial_folder=default_folder)
-            elif key == '__BUJ_Load_Flip_Stack__':
-                val.update(pad=pad(0, 0, 5, 0), target='__BUJ_Flip_Stage_Load__', disabled=True)
-            elif key == '__BUJ_Load_Unflip_Stack__':
-                val.update(pad=pad(0, 0, 0, 0), target='__BUJ_Unflip_Stage_Load__', disabled=True)
-            elif key == '__BUJ_Load_Mask__':
-                val.update(pad=pad(5, 0, 0, 0), target='__BUJ_Mask_Stage_Load__', disabled=True,
-                           enable_events=True)
-            elif key == '__BUJ_Load_FLS1__':
-                val.update(target='__BUJ_FLS1_Staging__', pad=((10, 0), (5, 0)),
-                           disabled=True)
-            elif key == '__BUJ_Load_FLS2__':
-                val.update(target='__BUJ_FLS2_Staging__', pad=((10, 0), (5, 0)),
-                           disabled=True)
-            # Button element
-            elif key == '__BUJ_Set_Img_Dir__':
-                val.update(pad=pad(5, 0, 0, 0),
-                           metadata={'State': 'Def'})
-            elif key == '__BUJ_Set_FLS__':
-                val.update(pad=((136, 0), (4, 5)), disabled=True,
-                           metadata={'State': 'Def'})
-            elif key == '__BUJ_Reset_FLS__':
-                val.update(pad=((5, 136), (4, 5)), disabled=True)
-            elif key == '__BUJ_Reset_Img_Dir__':
-                val.update(pad=pad(5, 0, 0, 0))
-            elif key == '__BUJ_Flip_Align__':
-                val.update(pad=pad(0, 0, 5, 0), disabled=True)
-            elif key == '__BUJ_Unflip_Align__':
-                val.update(pad=pad(0, 0, 0, 0), disabled=True)
-            elif key == '__BUJ_Adjust__':
-                val.update(pad=pad(80, 0, 0, 0), disabled=True,
-                           metadata={'Set': 'Quit Adjustment', 'Def': 'Adjust Image',
-                                     'State': 'Def'})
-            elif key == '__BUJ_Make_Mask__':
-                val.update(pad=pad(30, 5, 0, 0), disabled=True,
-                           metadata={'Set': 'Finish Mask', 'Def': 'Create New',
-                                     'State': 'Def'})
-            elif key == '__BUJ_Reset_Mask__':
-                val.update(visible=False)
-            elif key == '__BUJ_Elastic_Align__':
-                val.update(pad=pad(38, 0, 3, 0), disabled=True)
-            elif key == '__BUJ_View__':
-                val.update(pad=pad(145, 0, 10, 10), disabled=True,
-                           metadata={'Set': 'Hide Image', 'Def': 'View Image',
-                                     'State': 'Def'})
-            # Checkbox element
-            elif key == '__BUJ_filter__':
-                val.update(pad=pad(5, 0, 5, 5), default=True)
-            elif key == '__BUJ_LS_interp__':
-                val.update(pad=pad(200, 0, 4, 20), default=True)
-            elif key == '__BUJ_horizontal_flip__':
-                val.update(default=True,  pad=pad(45, 0, 0, 0), disabled=True)
-            # Combo element
-            elif key == '__BUJ_FLS_Combo__':
-                val.update(enable_events=True, default_value='Two', size=(5, 1),
-                           readonly=True, disabled=True, pad=((3, 0), (10, 0)),
-                           metadata={'Set': 'Two', 'Def': 'Two', 'State': 'Def'})
-            elif key == '__BUJ_TFS_Combo__':
-                val.update(enable_events=True, default_value='Unflip/Flip', size=(11, 1),
-                           readonly=True, disabled=True, pad=((0, 0), (10, 0)),
-                           metadata={'Set': 'Unflip/Flip', 'Def': 'Unflip/Flip',
-                                     'State': 'Def'})
-            elif key == '__BUJ_exp_transf__':
-                val.update(pad=pad(0, 10, 5, 10), size=(9, 1),
-                           default_value='Affine', readonly=True)
-            elif key == '__BUJ_LS_exp_transf__':
-                val.update(pad=pad(0, 10, 4, 10), size=(9, 1),
-                           default_value='Affine', readonly=True)
-            elif key == '__BUJ_reg_mode__':
-                val.update(pad=pad(0, 10, 19, 5), size=(11, 1),
-                           default_value='Accurate', enable_events=True, readonly=True)
-            elif key == '__BUJ_init_def__':
-                val.update(pad=pad(0, 10, 4, 5), size=(11, 1),
-                           default_value='Very Coarse', readonly=True)
-            elif key == '__BUJ_final_def__':
-                val.update(pad=pad(0, 10, 4, 5), size=(11, 1),
-                           default_value='Coarse', readonly=True)
-            elif key == '__BUJ_Mask_View__':
-                val.update(pad=pad(5, 0, 5, 0), visible=True,
-                           default_value='Unflip', enable_events=True, readonly=True)
-            # Graph element
-            elif key == '__BUJ_Graph__':
-                val.update(background_color='#808080', pad=pad(14, 14, 10, 0),
-                           enable_events=True)
-
-            # Image elements
-            elif key == '__BUJ_Unflip_Spinner__':
-                val.update(filename=background_img, background_color=theme_background, pad=((5, 5), (2, 0)),
-                           metadata={'Set': spinner_gif, 'Def': background_img,
-                                     'State': 'Def'})
-            elif key == '__BUJ_Flip_Spinner__':
-                val.update(filename=background_img, background_color=theme_background, pad=((5, 5), (5, 0)),
-                           metadata={'Set': spinner_gif, 'Def': background_img,
-                                     'State': 'Def'})
-            elif key == '__BUJ_Elastic_Spinner__':
-                val.update(filename=background_img, background_color=theme_background, pad=((5, 5), (5, 0)),
-                           metadata={'Set': spinner_gif, 'Def': background_img,
-                                     'State': 'Def'})
-            # Input field element
-            elif key == '__BUJ_FLS1__':
-                val.update(enable_events=True, use_readonly_for_disable=True,
-                           disabled=True, size=(24, 1), pad=((10, 0), (5, 0)),
-                           disabled_readonly_background_color='#A7A7A7',
-                           metadata={'Set': 'None', 'Def': 'None', 'State': 'Def'})
-            elif key == '__BUJ_FLS2__':
-                val.update(enable_events=True, use_readonly_for_disable=True,
-                           disabled=True, size=(24, 1), pad=((10, 0), (5, 0)),
-                           disabled_readonly_background_color='#A7A7A7',
-                           metadata={'Set': 'None', 'Def': 'None', 'State': 'Def'})
-            elif key == '__BUJ_Image_Dir_Path__':
-                val.update(size=(30, 1), enable_events=True,
-                           use_readonly_for_disable=True,
-                           disabled_readonly_background_color='#A7A7A7')
-            elif key == '__BUJ_LS_igb__':
-                val.update(**inp_d)
-            elif key == '__BUJ_LS_spso__':
-                val.update(**inp_d)
-            elif key == '__BUJ_LS_min_im__':
-                val.update(**inp_d)
-            elif key == '__BUJ_LS_max_im__':
-                val.update(**inp_d)
-            elif key == '__BUJ_LS_fds__':
-                val.update(**inp_d)
-            elif key == '__BUJ_LS_fdob__':
-                val.update(**inp_d)
-            elif key == '__BUJ_LS_cncr__':
-                val.update(**inp_d)
-            elif key == '__BUJ_LS_max_al_err__':
-                val.update(**inp_d)
-            elif key == '__BUJ_LS_inlier_rat__':
-                val.update(**inp_d)
-            elif key == '__BUJ_igb__':
-                val.update(**inp_d)
-            elif key == '__BUJ_spso__':
-                val.update(**inp_d)
-            elif key == '__BUJ_min_im__':
-                val.update(**inp_d)
-            elif key == '__BUJ_max_im__':
-                val.update(**inp_d)
-            elif key == '__BUJ_fds__':
-                val.update(**inp_d)
-            elif key == '__BUJ_fdob__':
-                val.update(**inp_d)
-            elif key == '__BUJ_cncr__':
-                val.update(**inp_d)
-            elif key == '__BUJ_max_al_err__':
-                val.update(**inp_d)
-            elif key == '__BUJ_inlier_rat__':
-                val.update(**inp_d)
-            elif key == '__BUJ_min_num_inlier__':
-                val.update(**inp_d)
-            elif key == '__BUJ_div_w__':
-                val.update(**inp_d)
-            elif key == '__BUJ_curl_w__':
-                val.update(**inp_d)
-            elif key == '__BUJ_land_w__':
-                val.update(**inp_d)
-            elif key == '__BUJ_img_w__':
-                val.update(**inp_d)
-            elif key == '__BUJ_cons_w__':
-                val.update(**inp_d)
-            elif key == '__BUJ_stop_thresh__':
-                val.update(**inp_d)
-            elif key == '__BUJ_transform_x__':
-                val.update(pad=(5, 5), size=(4, 1), disabled=True,
-                           use_readonly_for_disable=True,
-                           disabled_readonly_background_color='#A7A7A7')
-            elif key == '__BUJ_transform_y__':
-                val.update(pad=(5, 5), size=(4, 1), disabled=True,
-                           use_readonly_for_disable=True,
-                           disabled_readonly_background_color='#A7A7A7')
-            elif key == '__BUJ_transform_rot__':
-                val.update(pad=(5, 5), size=(4, 1), disabled=True,
-                           use_readonly_for_disable=True,
-                           disabled_readonly_background_color='#A7A7A7')
-
-            elif key == "__BUJ_Unflip_Mask_Inp__":
-                val.update(default_text='None', pad=((0, 0), (5, 0)), size=(18, 1),
-                           disabled=True,
-                           use_readonly_for_disable=True,
-                           disabled_readonly_background_color='#A7A7A7',
-                           metadata={'Set': 'None', 'Def': 'None',
-                                     'State': 'Def'})
-            elif key == "__BUJ_Flip_Mask_Inp__":
-                val.update(default_text='None', pad=((0, 0), (3, 0)), size=(18, 1),
-                           disabled=True,
-                           use_readonly_for_disable=True,
-                           disabled_readonly_background_color='#A7A7A7',
-                           metadata={'Set': 'None', 'Def': 'None',
-                                     'State': 'Def'})
-
-            elif key == '__BUJ_Unflip_Stack_Inp__':
-                val.update(default_text='None', size=(30, 1), pad=(0, 0), enable_events=True,
-                           background_color=self.DEF_BACKGROUND,
-                           metadata={'Set': 'None', 'Def': 'None',
-                                     'State': 'Def'})
-            elif key == '__BUJ_Flip_Stack_Inp__':
-                val.update(default_text='None', size=(30, 1), pad=(0, 0), enable_events=True,
-                           background_color=self.DEF_BACKGROUND,
-                           metadata={'Set': 'None', 'Def': 'None',
-                                     'State': 'Def'})
-            elif key == '__BUJ_Image1__':
-                val.update(default_text='None',
-                           size=(30, 1), pad=(0, 0), background_color=self.DEF_BACKGROUND, enable_events=True,
-                           metadata={'Set': 'None', 'Def': 'None',
-                                     'State': 'Def'})
-            elif key == '__BUJ_Image2__':
-                val.update(default_text='None', pad=(0, 0),
-                           size=(30, 1), background_color=self.DEF_BACKGROUND,
-                           metadata={'Set': 'None', 'Def': 'None',
-                                     'State': 'Def'})
-            elif key == '__BUJ_Stack__':
-                val.update(default_text='None', pad=pad(0, 0, 0, 5), enable_events=True,
-                           size=(30, 1), background_color=self.DEF_BACKGROUND,
-                           metadata={'Set': 'None', 'Def': 'None',
-                                     'State': 'Def'})
-
-            # List element:
-            elif key == '__BUJ_Image_Choice__':
-                val.update(pad=pad(5, 0, 10, 10), size=(10, 3), no_scrollbar=True,
-                           select_mode='LISTBOX_SELECT_MODE_SINGLE', disabled=True,
-                           default_values='Unflip LS', enable_events=True)
-            # Radio element
-            elif key == '__BUJ_unflip_reference__':
-                val.update(default=True, size=(12, 2), pad=pad(95, 0, 2, 2),
-                           enable_events=True, disabled=True)
-            elif key == '__BUJ_flip_reference__':
-                val.update(size=(12, 2), pad=pad(10, 0, 2, 2),
-                           enable_events=True, disabled=True)
-            # Slider element
-            elif key == '__BUJ_Stack_Slider__':
-                val.update(range=(0, 0), size=(32, 15), default_value=1,
-                           enable_events=True, resolution=1, orientation='h',
-                           disable_number_display=True, pad=((0, 0), (10, 0)),
-                           metadata={'slider_range': (0, 0)})
-            elif key == '__BUJ_img_subsf__':
-                val.update(range=(0, 7), size=(25, 12), pad=pad(30, 0, 4, 4), default_value=0,
-                           resolution=1, orientation='h', disable_number_display=False, visible=True)
-            # Text element
-            elif key == '__BUJ_FLS1_Text__':
-                val.update(pad=((5, 0), (5, 0)),
-                           metadata={'Both': 'Both FLS', 'Single': '1 FLS',
-                                     'Two': "Unflip FLS"})
-            elif key == '__BUJ_FLS2_Text__':
-                val.update(pad=((5, 0), (5, 0)),
-                           metadata={'Both': '', 'Single': '',
-                                     'Two': "Flip FLS"})
             return val
 
         def REC_style(key, val) -> Dict:
@@ -796,12 +348,8 @@ class WindowStyle(object):
         # Check if key a part of tab
         # For correct key, return its style
         key_style = dict(key=key)
-        if key.startswith("home") or key.startswith("__Fiji"):
+        if key.startswith("home"):
             key_style = home_style(key, key_style)
-        elif key.startswith("__LS"):
-            key_style = LS_style(key, key_style)
-        elif key.startswith("__BUJ"):
-            key_style = BUJ_style(key, key_style)
         elif key.startswith("__REC"):
             key_style = REC_style(key, key_style)
         return key_style
